move career.js to data folder
