+++ src/views/RecruitmentDetail.vue
@@ -47,7 +47,7 @@
 </template>
 
 <script>
-const jobs = require('../career');
+const jobs = require('../data/career');
 
 export default {
   data: () => ({
