+++ src/views/RecruitmentInfo.vue
@@ -79,7 +79,7 @@
 </template>
 
 <script>
-const jobs = require('../career');
+const jobs = require('../data/career');
 
 export default {
   data: () => ({
